requirements/base.txt
# base requirements

Django==3.2.14
django-braces==1.15.0
django-taggit==3.0.0
psycopg2-binary==2.9.10
django-oauth-toolkit==2.1.0
mistune==2.0.4
djangorestframework==3.13.1
redis==4.3.4
Pillow==11.3.0
django-nested-admin==3.4.0
django-cors-headers==3.13.0
whitenoise==6.2.0
django-allauth==0.51.0
faker==13.15.0
django-filter==22.1
jsonmodels==2.5.1
djangorestframework-simplejwt==5.2.0
sentry-sdk==1.7.2
django-s3-storage==0.13.7
django-crum==0.7.9
django-redis==5.2.0
django-guardian==3.0.3
setuptools==66.1.1